Finished commenting changes and added the file header reader to the file browser for the implementation of ticket #354
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/newfileheaders@567 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/io/file.cpp
@@ -1086,7 +1086,7 @@ namespace NumeRe
     // file type
     std::string NumeReDataFile::getVersionString()
     {
-        return toString(versionMajor) + "." + toString(versionMinor) + "." + toString(versionBuild);
+        return toString((int)versionMajor) + "." + toString((int)versionMinor) + "." + toString((int)versionBuild);
     }
